Added the fitted y's for fused lasso. Potential error though
WARNING: If you straight-up follow the primal dual relationship,
you get one less jump. That is, using \hat{beta} = y - t(D)%*%\hat{u}.
To resolve this, I actually use the u with one more jump that supposed,
so there is the right number of jumps in beta. This is wrong though
since this beta is over-regularized? Need to look into this.