Starting to plot the intervals themselves
Warning: Something suspicious is going on? In FL (CI_1jump) simulation,
in trial 44 for row 105 of param (5-height at 0.5), the confidence
interval does not contain the observed value? What's going on?

Also, for varying n, the two-sided CI has a LOT lower power than
the one-sided p-value. What's going on? We should change the CI to
return both the one and two-sided intervals maybe as an option?