Fixed the signs for ss. Statistical error before
We define the sign of a jump (for the purpose of segment tests) as
the sign of the fitted mean function, NOT the original sign detected
in the sequence of steps in fused lasso or binary segmentation.

TODO: I never incorporated the sign of the jump into my previous tests...
Lol. This was because all my previous signals were always right-positive...